Revert "Don't reference count DBus clients, exit core immediately when one of them request it"
It make more sense to keep it as it was. Un the case of a Dataengine, Canonical AppIndicator / KDE Plasmoid systray RFC, the daemon can run event if the client have not been openned or have been closed.

This reverts commit 66a2b8eede0045aafa14d2560e793089199accef.

+++ daemon/src/dbus/instance.cpp
@@ -34,11 +34,24 @@
 Instance::Instance (DBus::Connection& connection)
     : DBus::ObjectAdaptor (connection, "/org/sflphone/SFLphone/Instance")
 {
+    count = 0;
 }
 
 void
-Instance::Quit(void)
+Instance::Register (const int32_t& pid UNUSED,
+                    const std::string& name UNUSED)
 {
-    Manager::instance().terminate();
-    Manager::instance().getDbusManager()->exit();
+    count++;
+}
+
+
+void
+Instance::Unregister (const int32_t& pid UNUSED)
+{
+    count --;
+
+    if (count <= 0) {
+        Manager::instance().terminate();
+        Manager::instance().getDbusManager()->exit();
+    }
 }
